# packages/typespec/openapi/admin.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "app-template Admin API",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/v1/accounts": {
      "get": {
        "operationId": "listAdminAccounts",
        "summary": "Lists customer accounts for Admin Console operators",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "explode": false
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "explode": false
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountListResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOperationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access is forbidden.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          }
        },
        "tags": ["admin-accounts"],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "createAdminAccount",
        "summary": "Creates a customer account through the Admin Console",
        "parameters": [],
        "responses": {
          "201": {
            "description": "The request has succeeded and a new resource has been created as a result.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOperationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access is forbidden.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state of the server.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOperationErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          }
        },
        "tags": ["admin-accounts"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/auth/contexts/{authContextId}/refresh": {
      "post": {
        "operationId": "refreshAdminOperatorSession",
        "summary": "Refreshes the Admin operator access token for one auth context",
        "parameters": [
          {
            "name": "authContextId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UlidId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOperatorContextRefreshResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access is forbidden.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          }
        },
        "tags": ["admin-auth"],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BearerContextRefreshRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/operator-setup/finish": {
      "post": {
        "operationId": "finishAdminOperatorSetup",
        "summary": "Finishes Admin operator setup and returns an operator session",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOperatorAuthSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOperationErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access is forbidden.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          }
        },
        "tags": ["admin-auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminOperatorSetupFinishRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/operator-setup/start": {
      "post": {
        "operationId": "startAdminOperatorSetup",
        "summary": "Starts Admin operator setup passkey registration",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyAddStartResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOperationErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access is forbidden.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          }
        },
        "tags": ["admin-auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminOperatorSetupStartRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/operator/current": {
      "get": {
        "operationId": "getCurrentAdminOperator",
        "summary": "Gets the current Admin operator",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCurrentOperatorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access is forbidden.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          }
        },
        "tags": ["admin-auth"],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountListResponse": {
        "type": "object",
        "required": ["requestId", "accounts"],
        "properties": {
          "requestId": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UlidId"
              }
            ],
            "description": "一覧取得 request に対応する canonical ULID の追跡 ID。"
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountSummary"
            },
            "description": "検索条件に一致した Product Account の要約一覧。"
          },
          "nextCursor": {
            "type": "string",
            "description": "次ページが存在するときだけ返す opaque cursor。"
          }
        },
        "description": "account 一覧 API の response。requestId は監査・追跡に使用し、accounts は account concept read model として返す。"
      },
...
      "AuthFailureResponse": {
        "type": "object",
        "required": ["requestId", "error"],
        "properties": {
          "requestId": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UlidId"
              }
            ],
            "description": "request に対応する canonical ULID の追跡 ID。"
          },
          "error": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthFailureClassification"
              }
            ],
            "description": "client が fail-close 処理を選ぶための normalized failure classification。"
          }
        },
        "description": "認証・認可 failure の共通 response。requestId と normalized error だけを返し、token や Cookie 値は含めない。"
      },
      "AuthOperationErrorResponse": {
        "type": "object",
        "required": ["requestId", "error"],
        "properties": {
          "requestId": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UlidId"
              }
            ],
            "description": "request に対応する canonical ULID の追跡 ID。"
          },
          "error": {
            "type": "string",
            "description": "client が表示・分岐に使う error code。secret や token 値は含めない。"
          }
        },
        "description": "入力不正や操作競合など、認証 credential 検証以外の operation error response。"
      },
...
      "CreateAccountResponse": {
        "type": "object",
        "required": ["requestId", "account", "auditEventId"],
        "properties": {
          "requestId": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UlidId"
              }
            ],
            "description": "作成 request に対応する canonical ULID の追跡 ID。"
          },
          "account": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountSummary"
              }
            ],
            "description": "作成された Product Account の read model。"
          },
          "auditEventId": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UlidId"
              }
            ],
            "description": "Account 作成 intent/outcome を記録する Admin audit event の canonical ULID。"
          }
        },
        "description": "account 作成 API の response。作成された Product Account と、対応する audit event の相関 ID だけを返す。"
      },
...
      "UlidId": {
        "type": "string",
        "description": "Canonical ULID string used for auth-owned resource and correlation identifiers."
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT 形式のアクセストークンを使用した Bearer 認証。Authorization ヘッダーに `Bearer <JWT>` を付与すること。トークンは短命の JWT であり、subject・sessionID・iat・exp を含む。"
      }
    }
  },
  "servers": [
    {
      "url": "https://admin.app-template.example",
      "description": "Admin API server",
      "variables": {}
    }
  ]
}